backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
pdf2image==1.16.3
pillow==10.1.0
sqlalchemy==2.0.23
pydantic==2.5.2
requests==2.31.0
python-dateutil==2.8.2
aiofiles==23.2.1
